Fixed  arity problem with left-associative tags
The problem was that smt constraints such as "(assert (or/and x))" resulted in an error
as there was an assertion that the amount of arguments must at least be 2.

This assertion is removed. The solution is just to evaluate the argument, changing
(assert (or/and x)) to (assert (x)). This seems to be a reasonable assumption.